backend/pyproject.toml
[project]
name = "prompt-optimizer-backend"
version = "0.1.0"
description = "FastAPI backend for the Prompt Optimizer web app"
requires-python = ">=3.10,<3.13"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.34.0",
    "python-multipart>=0.0.20",
    "anthropic>=0.51.0",
    "chromadb>=1.0.0",
    "llama-index>=0.12.35",
    "llama-index-vector-stores-chroma>=0.4.1",
    "llama-index-embeddings-huggingface>=0.5.0",
    "sentence-transformers>=3.0.0",
    "python-dotenv>=1.1.0",
    "pyyaml>=6.0",
    "sse-starlette>=2.0.0",
    "httpx>=0.27.0",
    "google-genai>=1.0.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["app"]
